Use a heap for Next for merges, and
pre-compute if there's many postings on the
unset path.

Add posting lookup benchmarks

Signed-off-by: Brian Brazil <brian.brazil@robustperception.io>

Fixes https://github.com/prometheus/prometheus/issues/5173

FYI @jacksontj @bwplotka