@tomwilkie 2019-02-22
We should talk: https://github.com/cortexproject/cortex/blob/master/pkg/querier/batch/merge.go#L11

;-)